Move AbstractMapFragmentWithControls.kt to tasks package
